+++ frontend/src/App.js
@@ -1,47 +1,21 @@
 import React, { useState } from 'react';
 import logo from './logo.svg';
 import './App.css';
+import { useEffect } from 'react';
+import { BrowserRouter as Router, Routes, Route } from 'react-router-dom';
+import Login from './components/login';
+import Home from './components/home';
+
 
 function App() {
-  const [username, setUsername] = useState('');
-  const [password, setPassword] = useState('');
-
-  const handleLogin = (e) => {
-    e.preventDefault();
-    // Add your login logic here
-    console.log('Logging in with:', { username, password });
-  };
-
-  return (
-    <div className="App">
-      <header className="App-header">
-        <h1>Login Page</h1>
-        <form onSubmit={handleLogin} className="login-form">
-          <div>
-            <label htmlFor="username">Username:</label>
-            <input
-              type="text"
-              id="username"
-              value={username}
-              onChange={(e) => setUsername(e.target.value)}
-              required
-            />
-          </div>
-          <div>
-            <label htmlFor="password">Password:</label>
-            <input
-              type="password"
-              id="password"
-              value={password}
-              onChange={(e) => setPassword(e.target.value)}
-              required
-            />
-          </div>
-          <button type="submit">Login</button>
-        </form>
-      </header>
-    </div>
-  );
+return (
+    <Router>
+      <Routes>
+        <Route path="/" element={<Login />} />
+        <Route path="/home" element={<Home />} />
+      </Routes>
+    </Router>
+ );
 }
 
 export default App;
