+++ frontend/src/components/home.js
@@ -1,0 +1,10 @@
+import React, { useState } from 'react';
+// import logo from './logo.svg';
+// import './App.css';
+
+
+function Home() {
+
+}
+
+export default Home;
