+++ frontend/src/components/login.js
@@ -1,0 +1,48 @@
+import React, { useState } from 'react';
+// import logo from './logo.svg';
+import './login.css';
+
+function login() {
+  const [username, setUsername] = useState('');
+  const [password, setPassword] = useState('');
+
+  const handleLogin = (e) => {
+    e.preventDefault();
+    // Add your login logic here
+    console.log('Logging in with:', { username, password });
+  };
+
+  return (
+    <div className="App">
+      <header className="App-header">
+        <h1>Login Page</h1>
+        <form onSubmit={handleLogin} className="login-form">
+          <div>
+            <label htmlFor="username">Username:</label>
+            <input
+              type="text"
+              id="username"
+              value={username}
+              onChange={(e) => setUsername(e.target.value)}
+              required
+            />
+          </div>
+          <div>
+            <label htmlFor="password">Password:</label>
+            <input
+              type="password"
+              id="password"
+              value={password}
+              onChange={(e) => setPassword(e.target.value)}
+              required
+            />
+          </div>
+          <button type="submit">Login</button>
+          <button onClick={() => navigate("/home")} style={styles.signupButton}> Sign up</button>
+        </form>
+      </header>
+    </div>
+  );
+}
+
+export default login;
